course_assistant/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
streamlit>=1.30.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
python-dotenv>=1.0.0
langchain-google-genai>=1.0.0
